% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{as_flextable.tstest.escc}
\alias{as_flextable.tstest.escc}
\title{Transform a test object into flextable}
\usage{
\method{as_flextable}{tstest.escc}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  include.decision = FALSE,
  use.symbols = TRUE,
  table.caption = x$test_name,
  footnote.reference = FALSE,
  ...
)
}
\arguments{
\item{x}{an object of which inherits a \dQuote{tstest.escc} class.}

\item{digits}{integer, used for number formatting. Optionally, to avoid
scientific notation, set \sQuote{options(scipen=999)}.}

\item{signif.stars}{logical. If TRUE, \sQuote{significance stars} are printed.}

\item{include.decision}{prints out whether to reject the NULL at the 5\% level
of significance.}

\item{use.symbols}{for tests which either have parameters for which the
latex symbols were included in the calling function or for which the tests
generate values which can be represented as latex symbols, then these will
be generated.}

\item{table.caption}{an optional string for the table caption.}

\item{footnote.reference}{whether to include the reference paper of the test
in the footnotes.}

\item{...}{not currently used. The returned object can be manipulated further
using flextable.}
}
\value{
A flextable object.
}
\description{
Transforms a \dQuote{tstest.escc} object into a flextable
with options on symbolic representation.
}
