% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{bread.cgarch.estimate}
\alias{bread.cgarch.estimate}
\alias{bread.dcc.estimate}
\title{Bread Method}
\usage{
\method{bread}{cgarch.estimate}(x, ...)

\method{bread}{dcc.estimate}(x, ...)
}
\arguments{
\item{x}{an object of class \dQuote{cgarch.estimate} or \dQuote{dcc.estimate}.}

\item{...}{not currently used.}
}
\value{
The inverse of the numerical hessian of the model.
}
\description{
Bread Method
}
\author{
Alexios Galanos
}
