% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cgarch_modelspec}
\alias{cgarch_modelspec}
\title{Generic Methods for the Copula GARCH model specification}
\usage{
cgarch_modelspec(object, ...)
}
\arguments{
\item{object}{a valid object.}

\item{...}{additional parameters specific to the method implemented.}
}
\value{
an object of some class.
}
\description{
Generic Methods for the Copula GARCH model specification
}
