% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.cgarch.estimate}
\alias{coef.cgarch.estimate}
\alias{coef}
\alias{coef.dcc.estimate}
\alias{coef.gogarch.estimate}
\title{Extract Model Coefficients}
\usage{
\method{coef}{cgarch.estimate}(object, ...)

\method{coef}{dcc.estimate}(object, ...)

\method{coef}{gogarch.estimate}(object, ...)
}
\arguments{
\item{object}{an estimated object from one of the models in the package.}

\item{...}{none}
}
\value{
A numeric named vector of estimated coefficients.
}
\description{
Extract the estimated coefficients of a model.
}
\author{
Alexios Galanos
}
