% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.dcc.estimate}
\alias{plot.dcc.estimate}
\alias{plot.cgarch.estimate}
\title{Dynamic Correlation Model Plots}
\usage{
\method{plot}{dcc.estimate}(
  x,
  y = NULL,
  series = NULL,
  index_format = "\%Y",
  labels = FALSE,
  cex_labels = 0.8,
  ...
)

\method{plot}{cgarch.estimate}(
  x,
  y = NULL,
  series = NULL,
  index_format = "\%Y",
  labels = FALSE,
  cex_labels = 0.8,
  ...
)
}
\arguments{
\item{x}{an object of class \dQuote{dcc.estimate} or \dQuote{cgarch.estimate}.}

\item{y}{not used}

\item{series}{for the constant correlation a vector of length 2 indicating the
series numbers to use for the pairwise correlation plot. For the dynamic
correlation model, if NULL will include all series, else a vector of integers
of the series to include.}

\item{index_format}{for the dynamic correlation plot the x-axis label formatting.}

\item{labels}{whether to include y-axis labels. For a large number of series
it is best to leave this as FALSE.}

\item{cex_labels}{the shrink factor for the y-axis labels if included.}

\item{...}{additional parameters passed to the plotting functions.}
}
\value{
plots of the correlation.
}
\description{
Dynamic Correlation Model Plots
}
