% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{escc_test}
\alias{escc_test}
\title{Engle-Sheppard Constant Correlation Test}
\usage{
escc_test(x, order = c(1, 1), lags = 1, constant = FALSE, ...)
}
\arguments{
\item{x}{an xts matrix of stationary data.}

\item{order}{the GARCH model order.}

\item{lags}{the number of lags to test for.}

\item{...}{none.}
}
\value{
An object of class \dQuote{tstest.escc} which has a print and
as_flextable method.
}
\description{
The constant correlation test of Engle and Sheppard (2001).
}
\details{
The whitened residuals from a constant correlation GARCH model are used to
test the hypothesis that they are IID using an artificial regression of the outer
products of tese residuals on a constant and their lagged outer products.
The null hypothesis of constant correlation can then be tested using a Wald-type
test on the constant and coefficients of the lagged regressors (which should
be zero under the null), with the test statistic asymptotically distributed
as Chi-squared with (lags + 1) degrees of freedom. In the returned table, the
parameter \dQuote{J} represents the test statistic.
}
\note{
The test makes use of parallel functionality for estimating the univariate GARCH
models and the user can control the number of threads used using
\code{\link[future]{plan}}.
}
\references{
\insertRef{Engle2001}{tsmarch}
}
