% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{estfun.cgarch.estimate}
\alias{estfun.cgarch.estimate}
\alias{estfun.dcc.estimate}
\title{Score Method}
\usage{
\method{estfun}{cgarch.estimate}(x, ...)

\method{estfun}{dcc.estimate}(x, ...)
}
\arguments{
\item{x}{an object of class \dQuote{cgarch.estimate} or \dQuote{dcc.estimate}.}

\item{...}{not currently used.}
}
\value{
The score matrix
}
\description{
Score Method
}
\details{
The function returns the numerical scores of the negative of the
log likelihood at the optimal solution. These are used in the calculation
of sandwich estimators.
}
\author{
Alexios Galanos
}
