% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{expected_shortfall}
\alias{expected_shortfall}
\alias{expected_shortfall.gogarch.predict}
\title{Expected Shortfall (ES) method for predicted and simulated objects}
\usage{
expected_shortfall(object, ...)

\method{expected_shortfall}{gogarch.predict}(object, weights = NULL, alpha = 0.05, abs_tol = 1e-06, ...)
}
\arguments{
\item{object}{an object from the predict or simulate methods.}

\item{...}{additional parameters passed to the \code{\link{tsconvolve}} method
for the GO-GARCH model.}

\item{weights}{a vector of weights of length equal to the number of series. If
NULL then an equal weight vector is used.}

\item{alpha}{the quantile level for the value at risk.}

\item{abs_tol}{the absolute tolerance used in the quadrature integration.}
}
\value{
an object of class \dQuote{tsmodel.predict} which is a matrix of dimensions
sim x h of the value at risk.
}
\description{
Expected Shortfall (ES) method for predicted and simulated objects
}
