% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.cgarch.estimate}
\alias{fitted.cgarch.estimate}
\alias{fitted}
\alias{fitted.cgarch.simulate}
\alias{fitted.cgarch.predict}
\alias{fitted.dcc.estimate}
\alias{fitted.dcc.simulate}
\alias{fitted.dcc.predict}
\alias{fitted.gogarch.estimate}
\alias{fitted.gogarch.predict}
\alias{fitted.gogarch.simulate}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{cgarch.estimate}(object, ...)

\method{fitted}{cgarch.simulate}(object, ...)

\method{fitted}{cgarch.predict}(object, ...)

\method{fitted}{dcc.estimate}(object, ...)

\method{fitted}{dcc.simulate}(object, ...)

\method{fitted}{dcc.predict}(object, ...)

\method{fitted}{gogarch.estimate}(object, ...)

\method{fitted}{gogarch.predict}(object, ...)

\method{fitted}{gogarch.simulate}(object, ...)
}
\arguments{
\item{object}{an object from one of the estimated, simulated or predicted
model classes in the package.}

\item{...}{not currently used.}
}
\value{
For an estimated object an \dQuote{xts} matrix of the fitted values (either
zero or a constant), else for simulated or predicted objects a horizon x n_series x n_draws
array of the distributional values. The array output includes attributes such as the
date index (for the predicted object), horizon, n_draws and series names.
}
\description{
Extract the fitted values from an object.
}
