% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{globalindices}
\alias{globalindices}
\title{Global Financial Indices Closing Value log Weekly Return}
\format{
\subsection{\code{globalindices}}{

A data frame with 1698 rows and 34 columns, with row names being the
YYYY-mm-dd date.
}
}
\source{
Yahoo Finance
}
\usage{
globalindices
}
\description{
A selection of the log weekly returns (Friday) of 34 global financial indices
from Yahoo Finance starting on 1991-12-13 and ending on 2024-06-21.
}
\keyword{datasets}
