% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogarch.R
\name{gogarch_modelspec}
\alias{gogarch_modelspec}
\title{GO-GARCH Model specification}
\usage{
gogarch_modelspec(
  y,
  distribution = c("norm", "nig", "gh"),
  model = "garch",
  order = c(1, 1),
  ica = "radical",
  components = NCOL(y),
  lambda_range = c(-5, 5),
  shape_range = c(0.1, 25),
  ...
)
}
\arguments{
\item{y}{an xts matrix of pre-filtered (residuals) stationary data.}

\item{distribution}{a choice of the normal, normal inverse gaussian or
generalized hyperbolic distribution (see details).}

\item{model}{the GARCH model to use for each factor.}

\item{order}{the GARCH model order.}

\item{ica}{the Independent Component Analysis algorithm. Current only the
RADICAL algorithm is available.}

\item{components}{the number of components to extract in the pre-whitening
phase,}

\item{lambda_range}{for the generalized hyperbolic distribution, the range of
the lambda parameter.}

\item{shape_range}{for the generalized hyperbolic distribution, the range of
the shape parameter (zeta).}

\item{...}{additional arguments passed to the \code{\link{radical}} function.}
}
\value{
an object of class \dQuote{gogarch.spec}.
}
\description{
GO-GARCH Model specification
}
