% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pit.gogarch.fft}
\alias{pit.gogarch.fft}
\title{Probability Integral Transform (PIT) for weighted FFT densities}
\usage{
\method{pit}{gogarch.fft}(object, actual, ...)
}
\arguments{
\item{object}{an object of class \dQuote{gogarch.fft} or \dQuote{gogarch.fftsim}.}

\item{actual}{a vector of realized values representing the weighted values of the
series for the period under consideration.}

\item{...}{not used.}
}
\value{
a matrix.
}
\description{
Probability Integral Transform (PIT) for weighted FFT densities
}
