% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.cgarch.estimate}
\alias{print.summary.cgarch.estimate}
\alias{print.summary.tsmarch.estimate}
\alias{print.summary.dcc.estimate}
\alias{print.summary.gogarch.estimate}
\title{Model Estimation Summary Print method}
\usage{
\method{print}{summary.cgarch.estimate}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)

\method{print}{summary.dcc.estimate}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)

\method{print}{summary.gogarch.estimate}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{an object of class \dQuote{summary.cgarch.estimate} or
\dQuote{summary.dcc.estimate}}

\item{digits}{integer, used for number formatting. Optionally, to avoid
scientific notation, set \sQuote{options(scipen=999)}.}

\item{signif.stars}{logical. If TRUE, ‘significance stars’ are printed for each coefficient.}

\item{...}{not currently used.}
}
\value{
Invisibly returns the original summary object.
}
\description{
Print method for class \dQuote{summary.cgarch.estimate} or
\dQuote{summary.dcc.estimate}
}
