% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{bread}
\alias{estfun}
\alias{newsimpact}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{sandwich}{\code{\link[sandwich:bread]{bread()}}, \code{\link[sandwich:estfun]{estfun()}}}

  \item{tsgarch}{\code{\link[tsgarch:newsimpact]{newsimpact()}}}
}}

