% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.cgarch.estimate}
\alias{summary.cgarch.estimate}
\alias{summary}
\alias{summary.dcc.estimate}
\alias{summary.gogarch.estimate}
\title{Model Estimation Summary}
\usage{
\method{summary}{cgarch.estimate}(object, vcov_type = "OP", ...)

\method{summary}{dcc.estimate}(object, vcov_type = "OP", ...)

\method{summary}{gogarch.estimate}(object, vcov_type = "OP", ...)
}
\arguments{
\item{object}{an object of class \dQuote{cgarch.estimate} or \dQuote{dcc.estimate}.}

\item{vcov_type}{the type of standard errors based on the vcov estimate (see \code{\link{vcov}}).}

\item{...}{not currently used.}
}
\value{
A list with summary information of class \dQuote{summary.cgarch.estimate} or
\dQuote{summary.dcc.estimate}.
}
\description{
Summary method for class \dQuote{cgarch.estimate} or \dQuote{dcc.estimate}.
}
