% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsaggregate.cgarch.estimate}
\alias{tsaggregate.cgarch.estimate}
\alias{tsaggregate}
\alias{tsaggregate.cgarch.simulate}
\alias{tsaggregate.cgarch.predict}
\alias{tsaggregate.dcc.estimate}
\alias{tsaggregate.dcc.simulate}
\alias{tsaggregate.dcc.predict}
\alias{tsaggregate.gogarch.estimate}
\alias{tsaggregate.gogarch.predict}
\alias{tsaggregate.gogarch.simulate}
\title{Weighted Moments Aggregation}
\usage{
\method{tsaggregate}{cgarch.estimate}(object, weights = NULL, ...)

\method{tsaggregate}{cgarch.simulate}(object, weights = NULL, ...)

\method{tsaggregate}{cgarch.predict}(object, weights = NULL, ...)

\method{tsaggregate}{dcc.estimate}(object, weights = NULL, ...)

\method{tsaggregate}{dcc.simulate}(object, weights = NULL, ...)

\method{tsaggregate}{dcc.predict}(object, weights = NULL, ...)

\method{tsaggregate}{gogarch.estimate}(object, weights = NULL, ...)

\method{tsaggregate}{gogarch.predict}(object, weights = NULL, ...)

\method{tsaggregate}{gogarch.simulate}(object, weights = NULL, ...)
}
\arguments{
\item{object}{an object of one of the model classes in the package.}

\item{weights}{an optional weighting vector. If NULL will use an equal weight vector.
It is also possible to pass a time varying weighting matrix with time along the
row dimension and equal to the number of time points or horizon.}

\item{...}{not currently used.}
}
\value{
A list with the weighted moments. For an estimated object class these
will be xts vectors whilst for the simulated and predicted class these will be
objects of class \dQuote{tsmodel.distribution} capturing the distributional
uncertainty and for which a default plot method exists.
}
\description{
Calculates and returns the weighted moments of the
estimated, simulated or predicted object.
}
