% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{value_at_risk}
\alias{value_at_risk}
\alias{value_at_risk.gogarch.predict}
\title{Value at Risk (VaR) method for predicted and simulated objects}
\usage{
value_at_risk(object, ...)

\method{value_at_risk}{gogarch.predict}(object, weights = NULL, alpha = 0.05, ...)
}
\arguments{
\item{object}{an object from the predict or simulate methods.}

\item{...}{additional parameters passed to the \code{\link{tsconvolve}} method
for the GO-GARCH model.}

\item{weights}{a vector of weights of length equal to the number of series. If
NULL then an equal weight vector is used.}

\item{alpha}{the quantile level for the value at risk.}
}
\value{
an object of class \dQuote{tsmodel.predict} which is a matrix of dimensions
sim x h of the value at risk.
}
\description{
Value at Risk (VaR) method for predicted and simulated objects
}
