% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dcc_modelspec.tsgarch.multi_estimate}
\alias{dcc_modelspec.tsgarch.multi_estimate}
\title{DCC GARCH model specification}
\usage{
\method{dcc_modelspec}{tsgarch.multi_estimate}(
  object,
  dynamics = c("constant", "dcc", "adcc"),
  dcc_order = c(1, 1),
  distribution = c("mvn", "mvt"),
  cond_mean = NULL,
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{tsgarch.multi_estimate}.}

\item{dynamics}{the type of correlation dynamics to use.}

\item{dcc_order}{the order of the dynamics in case of \dQuote{dcc} or \dQuote{adcc}
correlation models.}

\item{distribution}{the multivariate distribution. If using the \dQuote{mvt},
then the first stage univariate models should use the normal distribution.}

\item{cond_mean}{an optional matrix of the conditional mean for the series.
This is only available when every first stage model uses \code{arma = c(0,0)};
with ARMA first stage models the conditional mean is taken from the first
stage models automatically.}

\item{...}{not currently used.}
}
\value{
an object of class \dQuote{dcc.spec}.
}
\description{
DCC GARCH model specification
}
