% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.tsmarch.newsimpact}
\alias{plot.tsmarch.newsimpact}
\title{News Impact Surface Plot}
\usage{
\method{plot}{tsmarch.newsimpact}(x, y = NULL, ...)
}
\arguments{
\item{x}{an object of class \dQuote{tsmarch.newsimpact}.}

\item{y}{not used.}

\item{...}{not currently supported.}
}
\value{
a plot of the newsimpact surface
}
\description{
Plot method for newsimpact class.
}
